requirements.txt
streamlit==1.32.0
pandas==2.2.0
matplotlib==3.8.2
google-api-python-client==2.118.0
python-dotenv==1.0.1
japanize-matplotlib==1.1.3
seaborn==0.13.2 